LayoutTests/js/dom/resizable-array-buffer-serialization-expected.txt

Resizable ArrayBuffers should be serializable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS arrayBuffer.resizable is true
PASS arrayBuffer.byteLength is 36
PASS arrayBuffer.maxByteLength is 128
PASS cloned.resizable is true
PASS cloned.byteLength is 36
PASS cloned.maxByteLength is 128
PASS arrayBuffer.byteLength is 36
PASS arrayBuffer.maxByteLength is 128
PASS cloned.byteLength is 128
PASS cloned.maxByteLength is 128
PASS successfullyParsed is true

TEST COMPLETE
